refactoring(directory): relocate classes under the right directory as well as remove unnecessary ones.
