+++ server/juse/src/main/java/com/example/juse/helper/storage/LocalStorageService.java
@@ -1,4 +1,4 @@
-package com.example.juse.user.service;
+package com.example.juse.helper.storage;
 
 import com.example.juse.helper.utils.StringUtils;
 import lombok.extern.slf4j.Slf4j;
