+++ server/juse/src/main/java/com/example/juse/helper/storage/config/StorageConfig.java
@@ -1,7 +1,7 @@
 package com.example.juse.helper.storage.config;
 
-import com.example.juse.user.service.S3StorageService;
-import com.example.juse.user.service.StorageService;
+import com.example.juse.helper.storage.S3StorageService;
+import com.example.juse.helper.storage.StorageService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
