--- server/juse/src/main/java/com/example/juse/tester/BoardStub.java
@@ -1,18 +1,0 @@
-package com.example.juse.tester;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class BoardStub {
-
-    private String id;
-    private String url;
-    private String content;
-
-}
