+++ server/juse/src/test/java/com/example/juse/config/AppKeyConfigTest.java
@@ -1,4 +1,4 @@
-package com.example.juse;
+package com.example.juse.config;
 
 import com.example.juse.security.jwt.AppKeyConfig;
 import com.example.juse.security.jwt.SecretKeyHolder;
