+++ server/juse/src/test/java/com/example/juse/config/JpaTestConfig.java
@@ -1,4 +1,4 @@
-package com.example.juse;
+package com.example.juse.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
