+++ server/juse/src/test/java/com/example/juse/config/OAuthLocalProfileTest.java
@@ -1,4 +1,4 @@
-package com.example.juse.tester;
+package com.example.juse.config;
 
 import com.example.juse.security.config.OAuthProperties;
 import org.junit.jupiter.api.Test;
