+++ server/juse/src/test/java/com/example/juse/config/TestDBInstance.java
@@ -1,4 +1,4 @@
-package com.example.juse;
+package com.example.juse.config;
 
 import com.example.juse.board.entity.Board;
 import com.example.juse.board.repository.BoardRepository;
